requirements.txt
python-telegram-bot==21.3
yt-dlp>=2025.1.6
python-dotenv==1.0.1
pydub==0.25.1
matplotlib==3.8.4
requests==2.32.3
opencv-python>=4.9.0
numpy>=1.26.0

openai-whisper
torch
